requirements.txt
python-dotenv
kaggle
matplotlib
numpy
ipython
nbconvert
jupyter
path
scipy
tensorflow-directml-plugin
h5py
opencv-python
pandas
seaborn
scikit-image
scikit-learn
notebook